// src/app/core/models/quotation-status.enum.ts
export enum QuotationStatus {
    CartSaved = 'Cart_saved',
    QuoteRequested = 'Quote_requested',
    QuoteCreated = 'Quote_created',
    OrderPlaced = 'Order_placed',
    OrderInProduction = 'Order_in_production',
    QuoteExpired = 'Quote_expired',
    OrderCancelled = 'Order_cancelled',
    OrderDelivered = 'Order_delivered',
    SaveAndExitCart = "Save_and_exit_cart",
    FinaliseQuote = "Finalise_quote",
    InReview = "In_review",
    SaveAndExitQuote = "Save_and_Exit_quote",
    PaymentPending ="Payment_pending",
    QuoteCancel = "Quote_Cancel"
}